requirements_expr_all.txt
astropy==7.0.1
CairoSVG==2.7.1
densratio==0.3.0
easydict==1.13
lightgbm==4.6.0
lingam==1.9.1
matlab==0.1
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
pygam==0.9.1
rpy2==3.5.17
scikit_learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sklearn_pandas==2.2.0
tqdm==4.67.1
ucimlrepo==0.0.7
